// src/user-profile/dto/user-onboarding-progress.dto.ts
export enum USER_ONBOARDING_STAGES {
  SERVICE_SELECTION = 'serviceSelection',
  SERVICE_SETUP = 'serviceSetup',
  PROFILE_SETUP = 'profileSetup',
  SAFETY_QUIZ = 'safetyQuiz',
  SUBSCRIPTION = 'subscription',
}

export type SubFieldProgress = {
  complete: boolean;
  message?: string;
};

export enum SERVICE_SELECTION_SUBLIST {
  PROVIDER_SERVICE = 'PROVIDER_SERVICE',
}

export enum SERVICE_SETUP_SUBLIST {
  PROVIDER_PET_PREFERANCE = 'PROVIDER_PET_PREFERANCE',
  CANCELLATION_POLICY = 'CANCELLATION_POLICY',
  HOME_ATTRIBUTES = 'HOME_ATTRIBUTES',
  SERVICE_RATES = 'SERVICE_RATES',
  AVAILABILITY = 'AVAILABILITY',
}

export enum PROFILE_SETUP_SUBLIST {
  BASIC_INFO = 'BASIC_INFO',
  DETAILS = 'DETAILS',
  CONTACT = 'CONTACT',
  GALLERY = 'GALLERY',
  PET_MANAGEMENT = 'PET_MANAGEMENT',
}

export enum SAFETY_QUIZ_SUBLIST {
  COMPLETE_QUIZ = 'COMPLETE_QUIZ',
}

export enum SUBSCRIPTION_SUBLIST {
  SUBSCRIBED = 'SUBSCRIBED',
}